First commit
The Python module `datasets.py` provides a way to load the ics.uci.edu mushroom dataset.

The `utils.py` module provides some utility functions (shuffle, split) for working with datasets.

The `logistic.py` module trains a LogisticRegression (from scikit-learn) for the mushroom dataset.

The above dataset is included in the repository for convenience.