logistic.tmc_shapley: do not use TQDM exponential average
The TQDM progress bar of the outer loop was using exponential averaging (`smoothing=0.3` by default).
This doesn't match the prior information we have on the average time of each iteration, which is instead ~uniform and doesn't depend on time.
The estimated ETA given by TQDM is now way more accurate.